Fixed remaining build errors with OsvrAndroidRenderer implementation
